requirements.txt
boto3==1.35.37
botocore==1.35.37
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
taskipy==1.13.0
task==0.2.5
flask==2.0.3